- further fixes in semi-integration tests
